Added some crude reconnection logic + code actually working

+++ src/main/java/io/subnano/jvmmonitor/GcEvent.java
@@ -14,6 +14,6 @@ public interface GcEvent extends JvmEvent {
 
     String collector();
 
-    long pauseTime();
+    float pauseTime();
 
 }
